AVIBACKCOPIA2/ai-service/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
h11==0.16.0
idna==3.11
joblib==1.5.3
numpy>=1.26.0
pandas>=2.1.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
scikit-learn>=1.3.0
scipy>=1.11.0
six==1.17.0
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0
starlette==0.52.1
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata>=2023.3
uvicorn==0.41.0
xgboost>=2.0.0
shap>=0.44.0
azure-eventhub>=5.11.0
azure-core>=1.30.0
